Add collation support to sort function. Fixes #100
Support collation option in https://docs.mongodb.com/manual/reference/collation/.
This allows more flexible options when using the sort function on the cursor.
Some of the options of the collation spec have not been implemented because they do not map well to `localeCompare`.
These include; maxVariable, backwards, and normalization.