Correctly project and merge nested objects with missing keys.
The result of a projection may be incorrect for nested objects due to a bug that is sensitive to the order the keys
are processed and whether there are missing keys in some of the objects. This stems from not having correct positional
information when merge partial objects due to collapsed holes that are created by missing values.

The latest approach to address this is to preserve the holes created by missing values when resolving the object graph.
We achieve this with a special `MISSING` value which the merge function is aware of and handles properly.
The MISSING values serve as positional placeholders to show where to merge in objects from the next projected key.
At the end of the merge, all remaining MISSING values are filtered out to get the final object.

Related to issues #93, #105, and #106.