Properly validate plain object types for `isObject` function. Fixes #149
The use of constructor property for validating plain object types is not relyable.
What qualifies as a plain object has been clearly defined with this new implementation to allow for predictable and consistent behaviour.
Any value with the toString tag "[object Object]" and prototype value equal to "null" or the result of "Object.getPrototype({})" is considered a plain object.