Move collation to Options object.
Unlike other options derived in the processing of a query, the collation is provided externally.
Given that it may be specified on different operations to be used, it needs to be on the `Options` base type.
The `Config` object should be reserved for truly global configuration that affect types of collection lookup.

Also marked fields in `Options` and `Config` as readonly and added optional argument to `createConfig` to
ensure default keys are included when unspecified.