feature(ProcessingMode): Add new option to improve pipeline performance by smartly cloning objects.
Added `ProcessingMode` option which specifies how inputs and outputs should be treated. The four modes are;
`CLONE_ALL`, `CLONE_INPUT`, `CLONE_OUTPUT`, and `CLONE_OFF` with the latter set as the default and offers
the greatest speed up.

By default the `$out` operator always deep clones the objects so that intermediate results can be collected.