fix(util): provide consistent behaviour for isEqual and stringify on custom types.
Only apply referential equality for user-defined types. This means two different user-defined objects
with the same data will not be equal. This avoids having to deal with mutability issues of user objects.
Since mingo is a data library, we still provide deep-equality checks for native JS types only.
Updated stringify to be more robust for representing user-defined types in a consistent manner but do
not rely on it for equality checks.

Closes #386